rewrite delete folder endpoint and move it to dataset routers
